requirements.txt
alembic==1.14.1 
bcrypt==4.2.1
APScheduler==3.11.0
email-validator==2.2.0
fastapi==0.115.8
fastapi-mail==1.4.2
fastapi-pagination==0.12.34
fpdf==1.7.2
gunicorn==23.0.0
openpyxl==3.1.5
pandas==2.2.3
passlib==1.7.4
pip==25.0
PyMySQL==1.1.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.1
requests==2.32.3
setuptools==75.8.0
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
XlsxWriter==3.2.2
pydantic ==2.10.6 
pydantic-settings==2.7.1 
pytest ==8.3.4 
httpx ==0.28.1 
num2words ==0.5.10